requirements.txt
Flask
Flask-Migrate
connexion[swagger-ui]
sqlalchemy
flask-bcrypt
Flask-Mail
PyJWT
flask-marshmallow_sqlalcemy
